@@ -5,7 +5,7 @@ author: Jeroen Trappers <jeroen@appmire.be>
 homepage: https://github.com/jeroentrappers/flutter_bcrypt
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
